Remove SwiftUI `stateNullable` helper for `StringDesc`
The `StringDesc` protocol comes from projects that also incorporate moko-resources. For projects that use only moko-mvvm, this code was generating a compiler error:

```
ViewModelStateNullable.swift:99:60: Cannot find type 'StringDesc' in scope
```

Note that this particular helper didn't have a corresponding `state` version in `ViewModelState`. After this change, both `state` and `stateNullable` are consistent in terms of the methods they offer.